fixes all tests and benchmakrs
